Migrating to micro-builder partially to eliminate bugs
Can't use the generic function for updates because properties are changing name between the message and the database. It may be worth doing something about this in the micro-builder library to make the function even more flexible but for now I think it's fine like this. This uses the same code copied in and just tweaked so it should be relatively bug free from my experience.